+++ yolort/datasets/__init__.py
@@ -1,3 +1,3 @@
 # Copyright (c) 2021, Zhiqiang Wang. All Rights Reserved.
 from .datapipeline import DataPipeline
-from .detection_datamodule import DetectionDataModule, VOCDetectionDataModule, CocoDetectionDataModule
+from .datamodule import DetectionDataModule, VOCDetectionDataModule, CocoDetectionDataModule
